services/rapid-ocr/pyproject.toml
[project]
name = "rapidocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.3",
    "numpy>=1.26.4",
    "opencv-python>=4.10.0.84",
    "paddlepaddle-gpu>=2.6.2",
    "psutil>=6.0.0",
    "rapidocr-onnxruntime>=1.3.24",
    "rapidocr-paddle>=1.3.27",
    "requests>=2.32.3",
    "robyn>=0.61.2",
    "setuptools>=75.1.0",
    "torch>=2.4.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.2",
]
